+++ src/client/mod.rs
@@ -8,6 +8,7 @@ use crate::irc;
 use crate::login::LoginCredentials;
 use crate::message::commands::ServerMessage;
 use crate::message::IRCMessage;
+use crate::message::{IRCTags, PrivmsgMessage};
 use crate::transport::Transport;
 use std::collections::HashSet;
 use std::sync::Arc;
@@ -123,8 +124,48 @@ impl<T: Transport, L: LoginCredentials> TwitchIRCClient<T, L> {
     /// it will not be cut short or split into multiple messages (what happens is determined
     /// by the behaviour of the Twitch IRC server).
     pub async fn say(&self, channel_login: String, message: String) -> Result<(), Error<T, L>> {
-        // The prefixed "." prevents execution of commands
-        self.privmsg(channel_login, format!(". {}", message)).await
+        self.say_in_response(channel_login, message, None).await
+    }
+
+    /// Say a chat message in the given Twitch channel, but send it as a response to another message if `reply_to_id` is specified.
+    ///
+    /// Behaves the same as `say()` when `reply_to_id` is None, but tags the original message and it's sender if specified.
+    pub async fn say_in_response(
+        &self,
+        channel_login: String,
+        message: String,
+        reply_to_id: Option<String>,
+    ) -> Result<(), Error<T, L>> {
+        let mut tags = IRCTags::new();
+
+        if let Some(id) = reply_to_id {
+            tags.0.insert("reply-parent-msg-id".to_string(), Some(id));
+        }
+
+        let irc_message = IRCMessage::new(
+            tags,
+            None,
+            "PRIVMSG".to_string(),
+            vec![format!("#{}", channel_login), format!(". {}", message)], // The prefixed "." prevents commands from being executed
+        );
+        self.send_message(irc_message).await
+    }
+
+    /// Replies to a given `PrivmsgMessage`, tagging the original message and it's sender.
+    ///
+    /// Similarly to `say()`, this method strips the message of executing commands, but does not filter out messages which are too long.
+    /// Refer to `say()` for the exact behaviour.
+    pub async fn reply_to_privmsg(
+        &self,
+        message: String,
+        reply_to: &PrivmsgMessage,
+    ) -> Result<(), Error<T, L>> {
+        self.say_in_response(
+            reply_to.channel_login.clone(),
+            message,
+            Some(reply_to.message_id.clone()),
+        )
+        .await
     }
 
     /// Join the given Twitch channel (When a channel is joined, the client will receive messages